Test CI/CD

+++ src/upgradeable/MineFunCore.sol
@@ -26,8 +26,7 @@ abstract contract MineFunCore is MineFunAdmin, IMineFun {
     address public soloTokenAddress;
 
     // Upper limit to avoid a too high of a paywall
-    uint256 public constant upperLimitMinSoloHeldForTokenCreation =
-        1000000 ether;
+    uint256 public constant upperLimitMinSoloHeldForTokenCreation = 1000000 ether;
 
     function setSoloTokenAddress(address newToken) public onlyOwner {
         soloTokenAddress = newToken;
@@ -38,23 +37,17 @@ abstract contract MineFunCore is MineFunAdmin, IMineFun {
         string memory symbol,
         string memory CIDLink,
         string memory imageCID,
-        uint bondingTime,
+        uint256 bondingTime,
         bool proxyCreation,
-        uint timestampOverride,
-        uint blockNumberOverride,
-        uint soloRequiredToMine
+        uint256 timestampOverride,
+        uint256 blockNumberOverride,
+        uint256 soloRequiredToMine
     ) public payable override returns (address) {
-        require(
-            msg.value >= MINEDTOKEN_CREATION_FEE,
-            "Insufficient creation fee"
-        );
-        require(
-            bondingTime >= 1 minutes && bondingTime <= 7 days,
-            "Bonding time must be between 1 and 7 days"
-        );
+        require(msg.value >= MINEDTOKEN_CREATION_FEE, "Insufficient creation fee");
+        require(bondingTime >= 1 minutes && bondingTime <= 7 days, "Bonding time must be between 1 and 7 days");
 
-        uint timestamp;
-        uint blockNum;
+        uint256 timestamp;
+        uint256 blockNum;
 
         if (proxyCreation) {
             require(timestampOverride > 0, "Invalid timestamp");
@@ -67,36 +60,22 @@ abstract contract MineFunCore is MineFunAdmin, IMineFun {
         }
 
         // Generate a unique salt
-        bytes32 salt = keccak256(
-            abi.encodePacked(name, symbol, msg.sender, timestamp, blockNum)
-        );
+        bytes32 salt = keccak256(abi.encodePacked(name, symbol, msg.sender, timestamp, blockNum));
 
         // Get the contract bytecode
-        bytes memory bytecode = abi.encodePacked(
-            type(Token).creationCode,
-            abi.encode(name, symbol, INIT_SUPPLY)
-        );
+        bytes memory bytecode = abi.encodePacked(type(Token).creationCode, abi.encode(name, symbol, INIT_SUPPLY));
 
         address minedTokenAddress;
 
         // Deploy using CREATE2
         assembly {
-            minedTokenAddress := create2(
-                0,
-                add(bytecode, 32),
-                mload(bytecode),
-                salt
-            )
-            if iszero(extcodesize(minedTokenAddress)) {
-                revert(0, 0)
-            }
+            minedTokenAddress := create2(0, add(bytecode, 32), mload(bytecode), salt)
+            if iszero(extcodesize(minedTokenAddress)) { revert(0, 0) }
         }
 
         require(minedTokenAddress != address(0), "CREATE2 failed");
 
-        MinedToken storage newMinedToken = addressToMinedTokenMapping[
-            minedTokenAddress
-        ];
+        MinedToken storage newMinedToken = addressToMinedTokenMapping[minedTokenAddress];
         newMinedToken.name = name;
         newMinedToken.symbol = symbol;
         newMinedToken.metadataCID = CIDLink;
@@ -130,16 +109,11 @@ abstract contract MineFunCore is MineFunAdmin, IMineFun {
     }
 
     function mineToken(address minedTokenAddress) public payable override {
-        MinedToken storage listedToken = addressToMinedTokenMapping[
-            minedTokenAddress
-        ];
+        MinedToken storage listedToken = addressToMinedTokenMapping[minedTokenAddress];
         require(listedToken.tokenAddress != address(0), "Token not found");
         require(!listedToken.bonded, "Token already bonded");
 
-        require(
-            block.timestamp < listedToken.bondingDeadline,
-            "Bonding period expired"
-        );
+        require(block.timestamp < listedToken.bondingDeadline, "Bonding period expired");
 
         require(soloTokenAddress != address(0), "Solo token address not set");
 
@@ -151,23 +125,18 @@ abstract contract MineFunCore is MineFunAdmin, IMineFun {
         );
 
         require(
-            IERC20(minedTokenAddress).balanceOf(msg.sender) + TOKENS_PER_MINE <=
-                MAX_PER_WALLET,
+            IERC20(minedTokenAddress).balanceOf(msg.sender) + TOKENS_PER_MINE <= MAX_PER_WALLET,
             "Maximum mine per wallet reached"
         );
 
         require(msg.value == PRICE_PER_MINE, "Incorrect ETH amount sent");
 
-        uint ethForLiquidity = msg.value / 2;
-        uint ethForTeam = msg.value - ethForLiquidity;
+        uint256 ethForLiquidity = msg.value / 2;
+        uint256 ethForTeam = msg.value - ethForLiquidity;
         teamFunds[minedTokenAddress] += ethForTeam;
 
-        uint totalTokensAfterPurchase = listedToken.tokensBought +
-            TOKENS_PER_MINE;
-        require(
-            totalTokensAfterPurchase <= INIT_SUPPLY,
-            "Not enough tokens left"
-        );
+        uint256 totalTokensAfterPurchase = listedToken.tokensBought + TOKENS_PER_MINE;
+        require(totalTokensAfterPurchase <= INIT_SUPPLY, "Not enough tokens left");
 
         Token minedToken = Token(minedTokenAddress);
         listedToken.fundingRaised += ethForLiquidity;
@@ -180,13 +149,9 @@ abstract contract MineFunCore is MineFunAdmin, IMineFun {
 
         if (listedToken.fundingRaised >= MINEDTOKEN_FUNDING_GOAL) {
             _createLiquidityPool(minedTokenAddress);
-            uint remainingTokens = MAX_SUPPLY - INIT_SUPPLY;
+            uint256 remainingTokens = MAX_SUPPLY - INIT_SUPPLY;
             minedToken.mint(address(this), remainingTokens);
-            _provideLiquidity(
-                minedTokenAddress,
-                remainingTokens,
-                listedToken.fundingRaised
-            );
+            _provideLiquidity(minedTokenAddress, remainingTokens, listedToken.fundingRaised);
             listedToken.bonded = true;
             emit TokenBonded(minedTokenAddress, listedToken.fundingRaised);
 
@@ -195,13 +160,8 @@ abstract contract MineFunCore is MineFunAdmin, IMineFun {
         }
     }
 
-    function updateSoloRequiredToMine(
-        address minedTokenAddress,
-        uint newRequirement
-    ) external {
-        MinedToken storage listedToken = addressToMinedTokenMapping[
-            minedTokenAddress
-        ];
+    function updateSoloRequiredToMine(address minedTokenAddress, uint256 newRequirement) external {
+        MinedToken storage listedToken = addressToMinedTokenMapping[minedTokenAddress];
         require(listedToken.tokenAddress != address(0), "Token not found");
         require(msg.sender == listedToken.creatorAddress, "Not token creator");
 
@@ -214,9 +174,7 @@ abstract contract MineFunCore is MineFunAdmin, IMineFun {
      * @param memeTokenAddress Address of the token
      * @return Address of the created pair
      */
-    function _createLiquidityPool(
-        address memeTokenAddress
-    ) internal returns (address) {
+    function _createLiquidityPool(address memeTokenAddress) internal returns (address) {
         IUniswapV2Factory factory = IUniswapV2Factory(UNISWAP_V2_FACTORY);
 
         address pair = factory.createPair(memeTokenAddress, router.WETH());
@@ -231,21 +189,15 @@ abstract contract MineFunCore is MineFunAdmin, IMineFun {
      * @param ethAmount Amount of ETH to provide
      * @return Amount of liquidity tokens received
      */
-    function _provideLiquidity(
-        address minedTokenAddress,
-        uint tokenAmount,
-        uint ethAmount
-    ) internal returns (uint) {
+    function _provideLiquidity(address minedTokenAddress, uint256 tokenAmount, uint256 ethAmount)
+        internal
+        returns (uint256)
+    {
         Token minedToken = Token(minedTokenAddress);
         minedToken.approve(UNISWAP_V2_ROUTER, tokenAmount);
 
-        (, , uint liquidity) = router.addLiquidityETH{value: ethAmount}(
-            minedTokenAddress,
-            tokenAmount,
-            tokenAmount,
-            ethAmount,
-            address(this),
-            block.timestamp
+        (,, uint256 liquidity) = router.addLiquidityETH{value: ethAmount}(
+            minedTokenAddress, tokenAmount, tokenAmount, ethAmount, address(this), block.timestamp
         );
 
         emit LiquidityProvided(minedTokenAddress, tokenAmount, ethAmount);
@@ -257,27 +209,22 @@ abstract contract MineFunCore is MineFunAdmin, IMineFun {
      * @param minedTokenAddress Address of the token
      */
     function refundContributors(address minedTokenAddress) public override {
-        MinedToken storage listedToken = addressToMinedTokenMapping[
-            minedTokenAddress
-        ];
-        require(
-            block.timestamp > listedToken.bondingDeadline,
-            "Bonding deadline not reached"
-        );
+        MinedToken storage listedToken = addressToMinedTokenMapping[minedTokenAddress];
+        require(block.timestamp > listedToken.bondingDeadline, "Bonding deadline not reached");
         require(!listedToken.bonded, "Token bonded, no refunds available");
 
-        uint contribution = listedToken.contributions[msg.sender];
+        uint256 contribution = listedToken.contributions[msg.sender];
         require(contribution > 0, "No contribution found");
 
         Token minedToken = Token(minedTokenAddress);
         minedToken.launchToken();
-        uint userTokens = (contribution / PRICE_PER_MINE) * TOKENS_PER_MINE;
+        uint256 userTokens = (contribution / PRICE_PER_MINE) * TOKENS_PER_MINE;
 
         minedToken.burn(msg.sender, userTokens);
         listedToken.contributions[msg.sender] = 0;
 
-        uint refundAmount = contribution;
-        uint teamFundForToken = teamFunds[minedTokenAddress];
+        uint256 refundAmount = contribution;
+        uint256 teamFundForToken = teamFunds[minedTokenAddress];
         if (teamFundForToken > 0) {
             teamFunds[minedTokenAddress] -= contribution / 2; // Remove team's portion too
             refundAmount = contribution;
@@ -293,13 +240,11 @@ abstract contract MineFunCore is MineFunAdmin, IMineFun {
      * @param minedTokenAddress Address of the token
      */
     function retrieveTeamFunds(address minedTokenAddress) public override {
-        MinedToken storage listedToken = addressToMinedTokenMapping[
-            minedTokenAddress
-        ];
+        MinedToken storage listedToken = addressToMinedTokenMapping[minedTokenAddress];
         require(listedToken.bonded, "Token did not bond");
         require(msg.sender == teamWallet, "Not authorized");
 
-        uint amount = teamFunds[minedTokenAddress];
+        uint256 amount = teamFunds[minedTokenAddress];
         require(amount > 0, "No funds available");
 
         teamFunds[minedTokenAddress] = 0;
